+++ src/components/Common/SaveButton.jsx
@@ -5,6 +5,12 @@ import bxCamera from '@iconify-icons/bx/bx-camera';
 import { FaShareSquare } from 'react-icons/fa';
 import { FaLink } from 'react-icons/fa6';
 import { Color } from '@styles';
+import { encodeData } from '../Common/Utils';
+import { toast } from 'react-toastify';
+import { useApi } from '../../hooks/';
+import { useRecoilValue } from 'recoil';
+import { selectedMyTableContentsState, selectedFieldState } from '../../recoils/atoms';
+import html2canvas from 'html2canvas';
 
 const Btn = styled.button`
 	position: fixed;
@@ -114,14 +120,55 @@ const UrlBtn = styled.div.attrs({
 	}
 `;
 
-const SaveButton = ({ onClickURL, onClickCapture }) => {
+const SaveButton = ({ roadmapContentRef }) => {
+	const { serverApi } = useApi();
+
+	const selectedMyTableContents = useRecoilValue(selectedMyTableContentsState);
+	const selectedFieldData = useRecoilValue(selectedFieldState);
+
+	// URL Button Click 이벤트
+	const handleURLButtonClick = () => {
+		const myTableDataString = JSON.stringify(selectedMyTableContents);
+		const encodedMyTableData = encodeData(myTableDataString);
+
+		const selectedFieldDataString = JSON.stringify(selectedFieldData);
+		const encodedSelectedFieldData = encodeData(selectedFieldDataString);
+
+		const baseURL = serverApi.defaults.baseURL;
+		const newUrl = `${baseURL}/road-map?myTableData=${encodedMyTableData}&selectedFieldData=${encodedSelectedFieldData}`;
+		notify_url('주소가 복사되었습니다.');
+
+		navigator.clipboard.writeText(newUrl).catch((error) => console.log(error));
+	};
+
+	// 스크린샷 Button Click 이벤트
+	const handleCaptureButtonClick = () => {
+		if (roadmapContentRef.current) {
+			html2canvas(roadmapContentRef.current)
+				.then((canvas) => {
+					const link = document.createElement('a');
+					link.href = canvas.toDataURL('image/png');
+					link.download = 'roadmap.png';
+					link.click();
+				})
+				.catch((error) => {
+					console.error('Error capturing the element:', error);
+				});
+		} else {
+			console.error('Roadmap content is not available');
+		}
+		notify_url('스크린샷을 저장하였습니다.');
+	};
+
+	const notify_url = (text) => toast.success(text);
+
 	return (
 		<>
 			<Btn>
-				<UrlBtn data-tooltip-content="Copy URL" data-tooltip-id="url" onClick={onClickURL}>
+				<UrlBtn data-tooltip-content="Copy URL" data-tooltip-id="url" onClick={handleURLButtonClick}>
 					<FaLink style={{ pointerEvents: 'none' }} />
 				</UrlBtn>
-				<CaptureBtn data-tooltip-content="Save as PNG" data-tooltip-id="capture" onClick={onClickCapture}>
+				<CaptureBtn data-tooltip-content="Save as PNG" data-tooltip-id="capture" onClick={handleCaptureButtonClick}>
 					<Icon style={{ pointerEvents: 'none' }} icon={bxCamera} />
 				</CaptureBtn>
 				<DownloadBtn>
